Extract generated finalizers into generic helper functions.
